backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
httpx==0.24.1
sqlmodel==0.0.8
asyncpg==0.28.0
aiosqlite==0.19.0
python-jose==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.0
openai>=1.0.0
alembic==1.11.1
python-multipart==0.0.20
redis==5.0.1
aioredis==2.0.1
argon2-cffi==23.1.0
pydantic==1.10.12
